Refactor master.js request handlers
