+++ src/request_handlers/setPlayerPositionSubscriptionRequestHandler.js
@@ -1,0 +1,8 @@
+"use strict";
+module.exports = async function setPlayerPositionSubscriptionRequestHandler(message, request, link) {
+	if (message.data.player_position) {
+		this.subscribedControlLinks.add(link);
+	} else {
+		this.subscribedControlLinks.delete(link);
+	}
+};
