+++ src/request_handlers/startInstanceRequestHandler.js
@@ -1,0 +1,6 @@
+"use strict";
+const { libLink } = require("@clusterio/lib");
+
+module.exports = async function startInstanceRequestHandler(message, request, link) {
+	return libLink.messages.startInstance.send(link, message.data);
+};
